Remove schemas.go and move related methods to test scope
